Got Start/Password choice menu/BossLevelSelect Menu to work.
Needed to have a state machine for the boss level selection menu for proper functionality and UI of which level is being hovered. Also needed to separate the flashy animation for that level specifically into it's own script because of empty gameobjects. Created a second script for Button Navigation for the Start/Password menu scene due to lack of knowledge on getting the same script to work for both scenes without changing vital information. I have an idea, but want to test it later when more content is added into project.